:bug: Convert from `DECIMAL` to `DOUBLE`
`DOUBLE` type provide a representation of a number, but `DECIMAL` provides a representation of a string, so `DOUBLE` is more convenience

+++ src/models/Product.ts
@@ -27,7 +27,7 @@ export default class Product extends Model {
     color!: string;
   
   @Column({
-    type: DataType.DECIMAL(8, 2),
+    type: DataType.DOUBLE(5, 2),
     allowNull: false,
   })
     price!: number;
